Logger: enforce UTF-8 encoding
Ensure we write UTF-8 encoded text to the log file. That is a trivial one since Qt uses unicode encoding internally anyway, almost all editors support UTF-8, and we only have full control over the file (and only open it for writing only).

In addition, we also enforce UTF-8 encoding for console output. It is a bit trickier. On Linux and macOS UTF-8 is the default and most probably nothing we need to worry about. Window, on the other hand, (especially older versions) are more complicated. It may only use a Latin-1 type encoding, like windows-2151 on my machine, but probably has everything required for UTF-8 already installed. Therefore, we explicitly try to set the encoding used in the console to UTF-8. This works fine on my end and if not, we probably just have some falsely displayed characters on screen. So, nothing worse than in the current version. However, the log file should be nicely encoded on all platforms now