Preferences: add option to always show type labels
previously, these labels have always been shown. Even in case all notes in a pattern do correspond to a instrument of the current kit (in which case type labels are somewhat redundant).

This patch introduces an option stored in `Preferences` and accessible via the right-click popup menu in the sidebar of the pattern editor using which the visibility of the instrument type labels can be controlled. If "Always show type labels" is checked, they will always be shown. If unchecked, they will be hidden in case all notes correspond to an instrument of the current kit.

Per default this new option is unchecked/off. I am not a 100% sure whether this works fine. Further testing and development will tell